reger 3.0/requirements.txt
aiogram==3.5.0
virtualbox
appium-python-client
sqlite3